New feature: Increase body condition in starvation.
In the starvation model of Illius & O’Connor
(2000), the mean body condition stays the same
even when a portion of the cohort is removed.
This feature is currently disabled, but can easily
switched on in herbiv_herbivore.cpp.

Additionally, the class
GetStarvationMortalityIllius2000 has been renamed
to include O’Connor, too.

To use the mortality with individuals is now
forbidden.