Redo shift of body condition with starvation.
- Changes class GetStarvationIlliusOConnor2000.
- Default value of parameter
  shift_body_condition_for_starvation is true.

The old approach was to calculate some kind of new
arithmetic mean of body condition after a fraction
of the cohort died. That created greater shifts,
but it also “generated” body fat out of nothing.

The new approach is mathematically simpler and
logically sound: the total fat (kg/km²) stays the
same before and after mortality.