Use HabitatForage::inedible instead of static data
In order to return something with the [] operators if `FT_INEDIBLE` was
passed to them, some empty static ForageBase object was returned. Using
static data is not advisable because of complications with
multi-threading. Also the static object could be changed when returned
as non-const reference.

Therefore, a “dummy” member variable HabitatForage::inedible was added.