Replace Hft::dead_herbivore_threshold
The `dead_herbivore_threshold` was a per cohort. The new
Hft::minimum_density_threshold is per population.

• This makes it applicable to any kind of HerbivoreInterface
implementation, not just cohorts.
• It is also easier to prevent invalid values (threshold below
establishment).
• Moreover, the newly established cohorts of long-lived HFTs could drop
through background mortality over their life time below a per-cohort threshold, which is a hard-to-mitigate artefact. With a per-population threshold that is not an issue anymore.
• The cohorts don’t need to take care to not create birth cohorts below
the minimum.