Include Catch2 as single header
I need to switch back to an included `catch.hpp` file because Catch2
is not available in the Debian repositories. This would only make it
frustrating for other users to install Catch2 manually on their systems.
This way the integration with CTest must also be dropped because the
CMake functions provided by the Catch2 installation are no longer
available.