Put documentation under CC-BY-4.0
For scientific work, the documentation should not be under a copyleft
license. It makes it very difficult to incorporate the work in other
scientific works. For that reason I decided against the GFDL and the
CC-BY-SA. (The GFDL has the additional disadvantageis that the full
license text needs to be attached always and that it doesn’t work well
for images.)