Allow anabolism efficiency > catabolism efficiency
When I implemented that check, I must have misunderstood the concept a
bit. If anabolism is more efficient than catabolism, that doesn’t mean
that energy could be gained out of nothing. There is always a cost in
storing fat because with storing fat, overall the ratio of invested
energy to regained energy is the *product* of anabolism and catabolism
efficiency, and this product will always be below 1 because both
coefficients are below 1.